Refactor per PR review
* Move logo_image to class level variable

* Use inheritance for admin form
